feat: prefer Prism for code block highlighting
The built-in implementation of Prism makes it awkward to add plugins, so
we disable it in favour of including our own bundled version of
prism.

These plugins clean up whitespace, adds selectable matching parens and
adds previews for things like CSS colours.

The first line of the prism.js file has the settings used to generate
the files.