Script to Adjust Margin for Blogs with Banner
I've created a script that will dynamically adjust the blog pages so that they will have a margin the size of the banner at the top so that all the content is displayed properly. This stops when the banner becomes hidden. This has been added to two blog pages and may need to be added to more layouts upon further testing.