Adjusted Layout for og:image and added one to XT25
I adjusted Layout.astro so that you can hard-code a path for an og:image (which seems to appear properly within the pages <head>) as well as it's previous use of being dynamically created for blogs. This was used to then add an og:image for XT25, as well as a description.