Created NYC Fintech Pages
Created a NYC Fintech 25 page, which copied from the XT25 page, adapting it's layout. It needed to use an agenda and contact us form, so I also copied and changed the equivalents to now fit this page. Some new images were used, so they were added to the images folder.