Set up a Scroll Threshold for Blogs & Fixed Subpages with Headers
I changed it so the alert banner was positioned properly on mobile for all pages (for blogs, career posts & team pages the banner & header would cut into the text/images being displayed). However, my fix still didn't appear to work well for blogs - unsure why - causing a white margin at the top of the screen.

To combat this, I implemented a scroll threshold to help keep the page scrolled down the size of the banner, which isn't perfect but mostly works on all platforms, and allows users to see all of the title.