Removed duplicate h1's in multiple pages
For indexing purposes, there can only be a single h1 per page, for most blog posts since there was a title in h1 already, I had to adjust the rest to the rest of the subtitles to be h2 and below, For the careers page, each position listing had an h1, so I changed this. For the "Team" pages,  there were two h1's within the greetings that needed to be adjusted. Finally, an adjustment to prettierignore had to be made to avoid a formatting problem.